backend/requirements.txt
Django==3.0.7
six==1.15.0
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
drf-yasg==1.17.1
gunicorn>=20.0.4
Pillow>=6.1.0